Fix Cygwin build

+++ source/threads/event_auto_reset.cpp
@@ -125,7 +125,7 @@ public:
             throwex SystemException("Failed to unlock a mutex for the auto-reset event!", result);
         return signaled;
 #elif defined(_WIN32) || defined(_WIN64)
-        DWORD result = WaitForSingleObject(_event, std::max(1, (DWORD)timespan.milliseconds()));
+        DWORD result = WaitForSingleObject(_event, std::max((DWORD)1, (DWORD)timespan.milliseconds()));
         if ((result != WAIT_OBJECT_0) && (result != WAIT_TIMEOUT))
             throwex SystemException("Failed to try lock an auto-reset event for the given timeout!");
         return (result == WAIT_OBJECT_0);
