+++ source/threads/named_event_manual_reset.cpp
@@ -178,7 +178,7 @@ public:
             throwex SystemException("Failed to unlock a mutex for the named manual-reset event!", result);
         return signaled;
 #elif defined(_WIN32) || defined(_WIN64) || defined(__CYGWIN__)
-        DWORD result = WaitForSingleObject(_event, std::max(1, (DWORD)timespan.milliseconds()));
+        DWORD result = WaitForSingleObject(_event, std::max((DWORD)1, (DWORD)timespan.milliseconds()));
         if ((result != WAIT_OBJECT_0) && (result != WAIT_TIMEOUT))
             throwex SystemException("Failed to try lock a named manual-reset event for the given timeout!");
         return (result == WAIT_OBJECT_0);
